+++ src/ios/FirebaseAnalyticsPlugin.m
@@ -1,7 +1,7 @@
 #import "FirebaseAnalyticsPlugin.h"
 
 @import Firebase;
-
+@import AppTrackingTransparency;
 
 @implementation FirebaseAnalyticsPlugin
 
@@ -77,5 +77,74 @@
     CDVPluginResult *pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
     [self.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
 }
+
+- (void)requestTrackingAuthorization:(CDVInvokedUrlCommand *)command {
+    
+    bool showInformation = [[command.arguments objectAtIndex:0] boolValue];
+
+    if(showInformation) {
+        
+        NSString* title = [command.arguments objectAtIndex:1];
+        NSString* message = [command.arguments objectAtIndex:2];
+        NSString* buttonTitle = [command.arguments objectAtIndex:3];
+        
+        [self showPermissionInformationPopup:title :message :buttonTitle :^(UIAlertAction *action ) {
+            [self showTrackingAuthorizationPopup:command];
+        }];
+        
+    }
+    else {
+        [self showTrackingAuthorizationPopup:command];
+    }
+}
+
+- (void)showTrackingAuthorizationPopup:(CDVInvokedUrlCommand *)command {
+    
+    if (@available(iOS 14, *)) {
+        [ATTrackingManager requestTrackingAuthorizationWithCompletionHandler:^(ATTrackingManagerAuthorizationStatus status) {
+            BOOL result = false;
+            switch(status) {
+                case ATTrackingManagerAuthorizationStatusAuthorized: {
+                    result = true;
+                    break;
+                }
+                default: {
+                    result = false;
+                    break;
+                }
+            }
+            CDVPluginResult *pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsBool:result];
+            [self.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
+        }];
+    }
+    else {
+        CDVPluginResult *pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsBool:true];
+        [self.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
+    }
+}
+
+typedef void (^showPermissionInformationPopupHandler)(UIAlertAction*);
+- (void)showPermissionInformationPopup:
+(NSString *)title :
+(NSString *)message :
+(NSString *)buttonTitle :
+(showPermissionInformationPopupHandler)confirmationHandler
+{
+    
+    UIAlertController *alert = [UIAlertController
+                                alertControllerWithTitle:title
+                                message:message
+                                preferredStyle:UIAlertControllerStyleAlert];
+    
+    UIAlertAction *okAction = [UIAlertAction
+                               actionWithTitle:buttonTitle
+                               style:UIAlertActionStyleDefault
+                               handler:confirmationHandler];
+    
+    [alert addAction:okAction];
+    [self.viewController presentViewController:alert animated:YES completion:nil];
+}
+
+
 
 @end